// src/lib/api/types.d.ts
export interface operations {
// ...
  articleBySlug: {
    parameters: {
      query?: never
      header?: never
      path: {
        slug: string
      }
      cookie?: never
    }
    requestBody?: never
    responses: {
      /** @description OK */
      200: {
        headers: Record<string, unknown>
        content: {
          "application/json": {
            id: string
            /** @enum {string} */
            language: "id" | "en"
            title: string
            slug: string
            content: string
            excerpt: string
            metaTitle: string | null
            metaDescription: string | null
            /** @enum {string} */
            status: "published" | "draft" | "rejected" | "in_review"
            /** @enum {string} */
            visibility: "public" | "member"
            articleTranslationId: string
            featuredImage: string
            createdAt: string | null
            updatedAt: string | null
            topics: {
              id: string
              title: string
              slug: string
            }[]
            authors: {
              id: string
              name: string | null
              username: string
            }[]
            editors: {
              id: string
              name: string | null
            }[]
          } | null
        }
      }
    }
  }
// ...
}